// src/types/character.ts
export interface Character {
  _id: number;
  name: string;
  films: string[];
  shortFilms: string[];
  tvShows: string[];
  videoGames: string[];
  parkAttractions: string[];
  allies: string[];
  enemies: string[];
  imageUrl: string;
  sourceUrl: string;
  createdAt: string;
  updatedAt: string;
}